requirements.txt
streamlit
requests
beautifulsoup4
black
isort
psycopg2-binary
python-dotenv
pandas
psycopg2